CMakeLists.txt: use CMAKE_INSTALL_FULL_LIBDIR for proper absolute path support
The CMAKE_INSTALL_LIBDIR is not necessarily a
relative path (may be an absolute one, e.g. for splitting libraries
from dev headers in package managers). Concatenating install prefix and such absolute libdir with "/" would result in an invalid path. The `_FULL_` variables are provided by GNUInstallDirs and are intended to handle this correctly.